+++ src/Html.php
@@ -18,6 +18,12 @@ use Yiisoft\Html\Tag\Colgroup;
 use Yiisoft\Html\Tag\CustomTag;
 use Yiisoft\Html\Tag\Div;
 use Yiisoft\Html\Tag\Em;
+use Yiisoft\Html\Tag\H1;
+use Yiisoft\Html\Tag\H2;
+use Yiisoft\Html\Tag\H3;
+use Yiisoft\Html\Tag\H4;
+use Yiisoft\Html\Tag\H5;
+use Yiisoft\Html\Tag\H6;
 use Yiisoft\Html\Tag\I;
 use Yiisoft\Html\Tag\Img;
 use Yiisoft\Html\Tag\Input;
@@ -44,6 +50,7 @@ use Yiisoft\Html\Tag\Textarea;
 use Yiisoft\Html\Tag\Tfoot;
 use Yiisoft\Html\Tag\Th;
 use Yiisoft\Html\Tag\Thead;
+use Yiisoft\Html\Tag\Title;
 use Yiisoft\Html\Tag\Tr;
 use Yiisoft\Html\Tag\Ul;
 use Yiisoft\Html\Widget\CheckboxList\CheckboxList;
@@ -397,6 +404,23 @@ final class Html
     }
 
     /**
+     * Generates a {@see Title} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
+     */
+    public static function title($content = '', array $attributes = []): Title
+    {
+        $tag = Title::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
      * Generates a {@see Meta} tag.
      *
      * @param array $attributes The tag attributes in terms of name-value pairs.
@@ -666,10 +690,14 @@ final class Html
      *
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes $attributes
      */
-    public static function textInput(?string $name = null, $value = null): Input
+    public static function textInput(?string $name = null, $value = null, array $attributes = []): Input
     {
-        return Input::text($name, $value);
+        $tag = Input::text($name, $value);
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
     /**
@@ -679,10 +707,14 @@ final class Html
      *
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes $attributes
      */
-    public static function hiddenInput(?string $name = null, $value = null): Input
+    public static function hiddenInput(?string $name = null, $value = null, array $attributes = []): Input
     {
-        return Input::hidden($name, $value);
+        $tag = Input::hidden($name, $value);
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
     /**
@@ -692,10 +724,14 @@ final class Html
      *
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes $attributes
      */
-    public static function passwordInput(?string $name = null, $value = null): Input
+    public static function passwordInput(?string $name = null, $value = null, array $attributes = []): Input
     {
-        return Input::password($name, $value);
+        $tag = Input::password($name, $value);
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
     /**
@@ -709,10 +745,14 @@ final class Html
      *
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes $attributes
      */
-    public static function fileInput(?string $name = null, $value = null): Input
+    public static function fileInput(?string $name = null, $value = null, array $attributes = []): Input
     {
-        return Input::file($name, $value);
+        $tag = Input::file($name, $value);
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
     /**
@@ -722,10 +762,14 @@ final class Html
      *
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes $attributes
      */
-    public static function radio(?string $name = null, $value = null): Radio
+    public static function radio(?string $name = null, $value = null, array $attributes = []): Radio
     {
-        return Input::radio($name, $value);
+        $tag = Input::radio($name, $value);
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
     /**
@@ -735,10 +779,14 @@ final class Html
      *
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes $attributes
      */
-    public static function checkbox(?string $name = null, $value = null): Checkbox
+    public static function checkbox(?string $name = null, $value = null, array $attributes = []): Checkbox
     {
-        return Input::checkbox($name, $value);
+        $tag = Input::checkbox($name, $value);
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
     /**
@@ -746,8 +794,11 @@ final class Html
      *
      * @param string|null $name The input name.
      * @param string|null $value The input value.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes $attributes
      */
-    public static function textarea(?string $name = null, ?string $value = null): Textarea
+    public static function textarea(?string $name = null, ?string $value = null, array $attributes = []): Textarea
     {
         $tag = Textarea::tag();
         if ($name !== null) {
@@ -756,7 +807,7 @@ final class Html
         if (!empty($value)) {
             $tag = $tag->value($value);
         }
-        return $tag;
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
     /**
@@ -915,6 +966,108 @@ final class Html
     public static function i($content = '', array $attributes = []): I
     {
         $tag = I::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
+     * Generates a {@see H1} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
+     */
+    public static function h1($content = '', array $attributes = []): H1
+    {
+        $tag = H1::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
+     * Generates a {@see H2} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
+     */
+    public static function h2($content = '', array $attributes = []): H2
+    {
+        $tag = H2::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
+     * Generates a {@see H3} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
+     */
+    public static function h3($content = '', array $attributes = []): H3
+    {
+        $tag = H3::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
+     * Generates a {@see H4} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
+     */
+    public static function h4($content = '', array $attributes = []): H4
+    {
+        $tag = H4::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
+     * Generates a {@see H5} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
+     */
+    public static function h5($content = '', array $attributes = []): H5
+    {
+        $tag = H5::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
+     * Generates a {@see H6} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     *
+     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
+     */
+    public static function h6($content = '', array $attributes = []): H6
+    {
+        $tag = H6::tag();
         if (!empty($attributes)) {
             $tag = $tag->attributes($attributes);
         }