+++ src/Tag/H3.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Html\Tag;
+
+use Yiisoft\Html\Tag\Base\NormalTag;
+use Yiisoft\Html\Tag\Base\TagContentTrait;
+
+/**
+ * @link https://html.spec.whatwg.org/multipage/sections.html#the-h1,-h2,-h3,-h4,-h5,-and-h6-elements
+ */
+final class H3 extends NormalTag
+{
+    use TagContentTrait;
+
+    protected function getName(): string
+    {
+        return 'h3';
+    }
+}